Ammend "Competence view to activity.py To maintain consistency with models and schemas
